runnerTest/requirements.txt
pandas
tensorflow
numpy
scikit-learn
scipy
matplotlib
influxdb-client
